+++ examples/node/scripts/follow/follow.ts
@@ -37,6 +37,7 @@ async function main() {
     throw new Error(`Something went wrong`);
   }
 
+  console.log(`Waiting for the transaction to be indexed...`);
   await client.transaction.waitUntilComplete({ forTxId: followResultValue.txId });
 }