+++ examples/web/src/misc/index.ts
@@ -1,3 +1,4 @@
+export * from './LensClientInteroperability';
 export * from './MiscPage';
 export * from './UseApproveModule';
 export * from './UseClaimHandle';