+++ packages/api-bindings/src/lens/__helpers__/mutations/authentication.ts
@@ -1,0 +1,19 @@
+import { MockedResponse } from '@apollo/client/testing';
+
+import {
+  RevokeAuthenticationData,
+  RevokeAuthenticationDocument,
+  RevokeAuthenticationVariables,
+} from '../../graphql/generated';
+
+export function mockRevokeAuthenticationResponse(
+  variables: RevokeAuthenticationVariables,
+): MockedResponse<RevokeAuthenticationData> {
+  return {
+    request: {
+      query: RevokeAuthenticationDocument,
+      variables,
+    },
+    result: { data: { revokeAuthentication: null } },
+  };
+}