+++ packages/react/src/transactions/adapters/profiles/BlockProfilesGateway.ts
@@ -22,19 +22,24 @@ import { ChainType, Data, PromiseResult, success } from '@lens-protocol/shared-k
 import { v4 } from 'uuid';
 
 import { UnsignedProtocolCall } from '../../../wallet/adapters/ConcreteWallet';
+import { IProviderFactory } from '../../../wallet/adapters/IProviderFactory';
+import { AbstractContractCallGateway, ContractCallDetails } from '../AbstractContractCallGateway';
 import { ITransactionFactory } from '../ITransactionFactory';
-import { SelfFundedProtocolTransactionRequest } from '../SelfFundedProtocolTransactionRequest';
 import { handleRelayError } from '../relayer';
 
 export class BlockProfilesGateway
+  extends AbstractContractCallGateway<BlockProfilesRequest>
   implements
     IDelegatedTransactionGateway<BlockProfilesRequest>,
     ISignedOnChainGateway<BlockProfilesRequest>
 {
   constructor(
+    providerFactory: IProviderFactory,
     private readonly apolloClient: SafeApolloClient,
     private readonly transactionFactory: ITransactionFactory<BlockProfilesRequest>,
-  ) {}
+  ) {
+    super(providerFactory);
+  }
 
   async createDelegatedTransaction(
     request: BlockProfilesRequest,
@@ -66,7 +71,6 @@ export class BlockProfilesGateway
       id: result.id,
       request,
       typedData: omitTypename(result.typedData),
-      fallback: this.createRequestFallback(request, result),
     });
   }
 
@@ -83,13 +87,15 @@ export class BlockProfilesGateway
     });
 
     if (data.result.__typename === 'LensProfileManagerRelayError') {
-      const result = await this.createTypedData(request);
-      const fallback = this.createRequestFallback(request, result);
-
-      return handleRelayError(data.result, fallback);
+      return handleRelayError(data.result);
     }
 
     return success(data.result);
+  }
+
+  protected async createEncodedData(request: BlockProfilesRequest): Promise<ContractCallDetails> {
+    const result = await this.createTypedData(request);
+    return this.createSetBlockStatusCallDetails(result);
   }
 
   private async createTypedData(
@@ -112,10 +118,9 @@ export class BlockProfilesGateway
     return data.result;
   }
 
-  private createRequestFallback(
-    request: BlockProfilesRequest,
+  private createSetBlockStatusCallDetails(
     result: CreateBlockProfilesBroadcastItemResult,
-  ): SelfFundedProtocolTransactionRequest<BlockProfilesRequest> {
+  ): ContractCallDetails {
     const contract = lensHub(result.typedData.domain.verifyingContract);
     const encodedData = contract.interface.encodeFunctionData('setBlockStatus', [
       result.typedData.message.byProfileId,
@@ -123,7 +128,6 @@ export class BlockProfilesGateway
       result.typedData.message.blockStatus,
     ]);
     return {
-      ...request,
       contractAddress: result.typedData.domain.verifyingContract,
       encodedData: encodedData as Data,
     };